+++ src/components/VisualSearchDebugAction/VisualSearchDebugViewer/index.ts
@@ -1,0 +1,1 @@
+export * from './VisualSearchDebugViewer';
